pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Autonomous AI Research Agent that gathers information, analyzes sources, and generates structured research reports."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.45.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.8.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
